+++ metrics/.gitignore
@@ -5,5 +5,4 @@
 _config
 dist
 .DS_Store
-/node_modules
-**/node_modules/**
+node_modules
