--- metrics/index.ts
@@ -1,2 +1,0 @@
-export { initializeOtel } from './init';
-export { incrementOtelCounter, observeOtelGauge, recordOtelHistogram } from './src/instruments';
