+++ metrics/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "typings": "",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc -p tsconfig.build.json"
   },
   "repository": {
     "type": "git",
