+++ metrics/src/index.ts
@@ -1,0 +1,2 @@
+export { initializeOtel } from './init';
+export { incrementOtelCounter, observeOtelGauge, recordOtelHistogram, instrument } from './instruments';
