+++ metrics/src/instruments/counter.ts
@@ -1,5 +1,5 @@
 import type { Attributes, Counter, MetricOptions } from '@opentelemetry/api';
-import { acquireMeter } from '../meter';
+import { acquireMeter } from 'src/lib/meter';
 
 const counters = {};
